feat: delete user

+++ rules/src/routes/role-permission/remove.ts
@@ -1,18 +1,31 @@
 import express, { Request, Response } from 'express';
 import { RolePermission } from '../../models/role-permission';
-import { NotFoundError } from '@share-package/common';
+import {
+  ListPermission,
+  NotFoundError,
+  UserType,
+  requireAuth,
+  requirePermission,
+  requireType,
+} from '@share-package/common';
 import { RolePermissionDeletedPublisher } from '../../events/publishers/role-permission/role-permission-deleted-publisher';
 import { natsWrapper } from '../../nats-wrapper';
 const router = express.Router();
-router.delete('/rules/delete/:id', async (req: Request, res: Response) => {
-  const { id } = req.params;
-  const rolePer = await RolePermission.findById(id);
-  if (!rolePer) throw new NotFoundError('Role-Permission');
-  await RolePermission.deleteOne({ _id: id });
-  new RolePermissionDeletedPublisher(natsWrapper.client).publish({
-    id: rolePer.id,
-    version: rolePer.version,
-  });
-  res.status(203).send({ message: 'Remove permission success' });
-});
+router.delete(
+  '/rules/delete/:id',
+  requireAuth,
+  requireType([UserType.Manager]),
+  requirePermission(ListPermission.RolePermissionDelete),
+  async (req: Request, res: Response) => {
+    const { id } = req.params;
+    const rolePer = await RolePermission.findById(id);
+    if (!rolePer) throw new NotFoundError('Role-Permission');
+    await RolePermission.deleteOne({ _id: id });
+    new RolePermissionDeletedPublisher(natsWrapper.client).publish({
+      id: rolePer.id,
+      version: rolePer.version,
+    });
+    res.status(203).send({ message: 'Remove permission success' });
+  }
+);
 export { router as deleteRolePermissionRouter };
